backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.13.0
django-filter==25.2
django-cors-headers==4.5.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==37.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
joblib==1.5.2
kombu==5.5.4
numpy==2.3.3
oauthlib==3.3.1
openai==2.3.0
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
spicy==0.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
gunicorn==21.2.0
Pillow>=10.0.0
